remove splinter v1; replace with splinter v2; reorganize code
