+++ src/count.rs
