+++ src/partition_ops.rs
@@ -1,26 +1,10 @@
-use std::ops::Deref;
-
-use crate::splinterv2::{
-    PartitionRead, PartitionWrite, SplinterRefV2, SplinterV2,
+use crate::{
+    PartitionRead, PartitionWrite,
     codec::partition_ref::{NonRecursivePartitionRef, PartitionRef},
     level::Level,
     partition::Partition,
     traits::{Cut, Merge},
 };
-
-impl<B: Deref<Target = [u8]>> PartialEq<SplinterRefV2<B>> for SplinterV2 {
-    #[inline]
-    fn eq(&self, other: &SplinterRefV2<B>) -> bool {
-        self.0 == other.load_unchecked()
-    }
-}
-
-impl<B: Deref<Target = [u8]>> PartialEq<SplinterV2> for SplinterRefV2<B> {
-    #[inline]
-    fn eq(&self, other: &SplinterV2) -> bool {
-        other == self
-    }
-}
 
 impl<L: Level> PartialEq for Partition<L> {
     fn eq(&self, other: &Partition<L>) -> bool {
@@ -190,34 +174,6 @@ impl<L: Level> Cut<PartitionRef<'_, L>> for Partition<L> {
     }
 }
 
-impl Merge for SplinterV2 {
-    fn merge(&mut self, rhs: &Self) {
-        self.0.merge(&rhs.0)
-    }
-}
-
-impl<B: Deref<Target = [u8]>> Merge<SplinterRefV2<B>> for SplinterV2 {
-    fn merge(&mut self, rhs: &SplinterRefV2<B>) {
-        self.0.merge(&rhs.load_unchecked())
-    }
-}
-
-impl Cut for SplinterV2 {
-    type Out = Self;
-
-    fn cut(&mut self, rhs: &Self) -> Self::Out {
-        Self(self.0.cut(&rhs.0))
-    }
-}
-
-impl<B: Deref<Target = [u8]>> Cut<SplinterRefV2<B>> for SplinterV2 {
-    type Out = Self;
-
-    fn cut(&mut self, rhs: &SplinterRefV2<B>) -> Self::Out {
-        Self(self.0.cut(&rhs.load_unchecked()))
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::collections::HashSet;
@@ -227,26 +183,24 @@ mod tests {
     use quickcheck_macros::quickcheck;
 
     use crate::{
-        splinterv2::{
-            Optimizable, SplinterV2,
-            traits::{Cut, Merge},
-        },
-        testutil::mksplinterv2,
+        Optimizable, Splinter,
+        testutil::mksplinter,
+        traits::{Cut, Merge},
     };
 
     #[quickcheck]
     fn test_partitions_equality_quickcheck(values: Vec<u32>) -> TestResult {
-        let mut a = mksplinterv2(&values);
+        let mut a = mksplinter(&values);
         a.optimize();
-        let b = mksplinterv2(&values);
+        let b = mksplinter(&values);
         TestResult::from_bool(a == b)
     }
 
     #[quickcheck]
     fn test_partitions_equality_ref_quickcheck(values: Vec<u32>) -> TestResult {
-        let mut a = mksplinterv2(&values);
+        let mut a = mksplinter(&values);
         a.optimize();
-        let b = mksplinterv2(&values).encode_to_splinter_ref();
+        let b = mksplinter(&values).encode_to_splinter_ref();
         TestResult::from_bool(a == b)
     }
 
@@ -254,9 +208,9 @@ mod tests {
     fn test_partitions_equality_quickcheck_2(a: Vec<u32>, b: Vec<u32>) -> TestResult {
         let expected = itertools::equal(a.iter().sorted().dedup(), b.iter().sorted().dedup());
 
-        let mut a = mksplinterv2(&a);
+        let mut a = mksplinter(&a);
         a.optimize();
-        let b = mksplinterv2(&b);
+        let b = mksplinter(&b);
 
         TestResult::from_bool((a == b) == expected)
     }
@@ -265,38 +219,38 @@ mod tests {
     fn test_partitions_equality_ref_quickcheck_2(a: Vec<u32>, b: Vec<u32>) -> TestResult {
         let expected = itertools::equal(a.iter().sorted().dedup(), b.iter().sorted().dedup());
 
-        let mut a = mksplinterv2(&a);
+        let mut a = mksplinter(&a);
         a.optimize();
-        let b = mksplinterv2(&b).encode_to_splinter_ref();
+        let b = mksplinter(&b).encode_to_splinter_ref();
 
         TestResult::from_bool((a == b) == expected)
     }
 
     #[quickcheck]
     fn test_merge_quickcheck(optimize: bool, a: HashSet<u32>, b: HashSet<u32>) -> TestResult {
-        let mut merged: SplinterV2 = a.iter().copied().collect();
-        let other: SplinterV2 = b.iter().copied().collect();
+        let mut merged: Splinter = a.iter().copied().collect();
+        let other: Splinter = b.iter().copied().collect();
 
         if optimize {
             merged.optimize();
         }
 
-        let expected: SplinterV2 = a.union(&b).copied().collect();
+        let expected: Splinter = a.union(&b).copied().collect();
         merged.merge(&other);
         TestResult::from_bool(merged == expected)
     }
 
     #[quickcheck]
     fn test_cut_quickcheck(optimize: bool, a: HashSet<u32>, b: HashSet<u32>) -> TestResult {
-        let mut source: SplinterV2 = a.iter().copied().collect();
-        let other: SplinterV2 = b.iter().copied().collect();
+        let mut source: Splinter = a.iter().copied().collect();
+        let other: Splinter = b.iter().copied().collect();
 
         if optimize {
             source.optimize();
         }
 
-        let expected_intersection: SplinterV2 = a.intersection(&b).copied().collect();
-        let expected_remaining: SplinterV2 = a.difference(&b).copied().collect();
+        let expected_intersection: Splinter = a.intersection(&b).copied().collect();
+        let expected_remaining: Splinter = a.difference(&b).copied().collect();
 
         let actual_intersection = source.cut(&other);
 
@@ -307,29 +261,29 @@ mod tests {
 
     #[quickcheck]
     fn test_merge_ref_quickcheck(optimize: bool, a: HashSet<u32>, b: HashSet<u32>) -> TestResult {
-        let mut merged: SplinterV2 = a.iter().copied().collect();
-        let other_ref = SplinterV2::from_iter(b.clone()).encode_to_splinter_ref();
+        let mut merged: Splinter = a.iter().copied().collect();
+        let other_ref = Splinter::from_iter(b.clone()).encode_to_splinter_ref();
 
         if optimize {
             merged.optimize();
         }
 
-        let expected: SplinterV2 = a.union(&b).copied().collect();
+        let expected: Splinter = a.union(&b).copied().collect();
         merged.merge(&other_ref);
         TestResult::from_bool(merged == expected)
     }
 
     #[quickcheck]
     fn test_cut_ref_quickcheck(optimize: bool, a: HashSet<u32>, b: HashSet<u32>) -> TestResult {
-        let mut source: SplinterV2 = a.iter().copied().collect();
-        let other_ref = SplinterV2::from_iter(b.clone()).encode_to_splinter_ref();
+        let mut source: Splinter = a.iter().copied().collect();
+        let other_ref = Splinter::from_iter(b.clone()).encode_to_splinter_ref();
 
         if optimize {
             source.optimize();
         }
 
-        let expected_intersection: SplinterV2 = a.intersection(&b).copied().collect();
-        let expected_remaining: SplinterV2 = a.difference(&b).copied().collect();
+        let expected_intersection: Splinter = a.intersection(&b).copied().collect();
+        let expected_remaining: Splinter = a.difference(&b).copied().collect();
 
         let actual_intersection = source.cut(&other_ref);
